+++ NEWS.md
@@ -3,6 +3,7 @@
 * Decided to remove genetics-related functions to instead have them in a separate package, as the dependencies were getting out of hand
   * See https://github.com/lukepilling/gwasRtools
 * Add `annotate_textp()` function for use in ggplot2 plots for adding easy text annotation geoms
+* Moved some functions to be internal only to tidy up the namespace
 
 # lukesRlib 0.2.2